// frontend/src/codegen/model/developers-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Flathub API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import { Developer } from "./developer"

/**
 *
 * @export
 * @interface DevelopersResponse
 */
export interface DevelopersResponse {
  /**
   *
   * @type {Array<Developer>}
   * @memberof DevelopersResponse
   */
  developers: Array<Developer>
  /**
   *
   * @type {Array<Developer>}
   * @memberof DevelopersResponse
   */
  invites: Array<Developer>
}